refactor 6502 instructions and flags into modules
