+++ crates/rmachine-mos6502/src/flags.rs
@@ -1,0 +1,2 @@
+pub const CARRY: u8 = 0b0000_0001;
+pub const DECIMAL: u8 = 0b0000_1000;
